chore(lint): expand golangci-lint config and fix all findings
Upgrade golangci-lint 2.0.2 -> 2.12.2 (host + .custom-gcl.yml) and enable a
pragmatic linter set on top of the modelguard plugin: correctness (errcheck,
govet, staticcheck, protogetter, bodyclose, noctx, nil*, gosec, ...), modern Go
(copyloopvar, intrange, modernize), low-noise quality linters, plus a goimports
formatter, depguard rules and a gosec G115 exclusion.

Resolve all resulting findings (222 -> 0):
- bodyclose: close the telegram response body
- gosec G112: set ReadHeaderTimeout on the gateway HTTP server
- noctx: http.NewRequest -> NewRequestWithContext, net.Listen -> ListenConfig,
  thread context through SendToTelegram
- gocritic / unconvert / intrange / revive cleanups, two justified gosec G118 nolints
- rename stuttering types: OrderByInfo->Info, ServerInfoRepository->Repository,
  VerificationStatus->Status

+++ internal/pkg/mongox/pagination/pagination.go
@@ -121,7 +121,7 @@ func Find[T Identifiable](
 		}
 	}
 
-	findOpts := options.Find().SetSort(pgOpts.sort).SetLimit(int64(pgOpts.limit))
+	findOpts := options.Find().SetSort(pgOpts.sort).SetLimit(pgOpts.limit)
 	cursor, err := coll.Find(ctx, pgOpts.filter, findOpts)
 	if err != nil {
 		return nil, err
